desktopIntegration/requirements.txt
absl-py==2.2.2
attrs==25.3.0
cffi==1.17.1
contourpy==1.1.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fastdtw==0.3.4
flatbuffers==25.2.10
fonttools==4.57.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
kiwisolver==1.4.7
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
MouseInfo==0.1.3
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
scipy==1.10.1
six==1.17.0
sounddevice==0.5.1
zipp==3.20.2
